requirements.txt
pillow>=10.0.0
pyautogui>=0.9.53
keyboard>=0.13.5 